requirements.txt
fastapi
uvicorn
requests
transformers
sentence-transformers
langchain
google-cloud-vision
pytesseract
faiss-cpu
openpyxl
python-docx